mach-exynos: Fix choice order in Kconfig.local for P4NOTE LTE variants
With MACH_P4NOTELTE_USA_VZW first this seems to become the default and
this causes other p4note variants to be detected as p4noteltevzw (tested by
inserting info statements into the various sections). In particular this breaks
GPS on n801x. Changing the order so that MACH_P4NOTE_USA_ANY is first fixes this.

Change-Id: Id9740488471b6db5d0709fd963ba6fc46fa93a7e