Update patch set 1
Patch Set 1:

As it says, "Part 1 of the Great Graphics Stack Update." At this point it's clearly not mergeable, which is why he hasn't +1ed it, but potentially when he finishes the whole process the camera issue may be resolved, and it should bring a lot of benefit overall.

IMO, the point of sharing the kernel codebase across Exynos 4210 and 4412 is so that work like this can be undertaken which can potentially help fix a lot of issues and improve performance across the whole family of devices. If we split them, all that will happen is that one of them will fall behind over time or someone will be sitting around just cherry picking fixes across from one kernel to the other.

Patch-set: 1