Update patch set 1
Patch Set 1:

Camera on 4412 devices will need updates to work with this but should be fixable.

The solution of course there is to fix the camera HALs to work with this, not hold back our ability to ever fix hwcomposer and our derpy graphics subsystem.

This will force n80xx to finally go dig up sbrissen's cam fixes so it can move to the opensource cam HAL.  That should probably be top priority once n80xx goes live with nightlies.  (just need to fix up the unified device props patch and we can light up n80xx...)

Patch-set: 1