Update patch set 3
Patch Set 3:

Even if the chip does support distinguishing CDP and SDP from each other, your code here does not do this.  You took the enum that is associated with an SDP being detected and changed the current from 500 mA to 1500.  (BTW, the "750 mA output from a 500 mA USB port" feature is due to having a switchmode power supply.  If the battery is very low, there will be almost 750 mA output at 3.3 volts if there is 500 mA input at 5v)

Changing the AC value to 1900 mA is OK though - although it's questionable whether it'll actually do anything.  The cables for most Samsung tablets are fairly thin and likely to "brown out" a bit at high currents, and the SMB347 has a notorious tendency to reduce current when the input voltage starts dropping.  (it's especially noticeable on the Nexus 7)

Patch-set: 3