Update patch set 3
Patch Set 3: Code-Review+1

A tester already reported that the fuse problems remain after this. That's a shame. Which makes this just a cleanup effort. He also reported that Kconfig needs the proper default for FORCE_MAX_ZONEORDER. Which led me to find weirdness here:

https://gerrit.omnirom.org/#/c/4426/15/drivers/gpu/mali400/r3p2/mali/common/mali_kernel_mem_os.c

For non-Note8 devices, the FORCE_MAX_ZONEORDER is hardcoded in Mali as 12. It wasn't like that. Sounds like that 'if defined' should have been 'if !defined'. However, their config is at 11...

This new patchset should fix that.

Patch-set: 3