[Do not merge] smdk4412: Enables ARM errata 761320 for smdk4210 as well.
ARM_ERRATA_761320 is already active for smdk4412 and smdk4212, but
should be applicable to the Cortex A9 r2p7 cores in smdk4210 as well.

Do not merge, for testing only, seems to be causing restarts in some
apps. However it also does seem to resolve or diminish the fuse problem.

Change-Id: I7685119bd738dd7e43c322b3c4af0d47fecb2714