Update patch set 1
Patch Set 1:

Hmm...  Would be useful to see what this errata is.

One thing is:  It is not enabled in the old smdk4210 kernel.

It may be useful to check the latest and greatest Samsung OSRC drop for one of these devices (N7000 is probably the most updated) to see if Samsung turned that on at some point.

Patch-set: 1
Reviewer: Gerrit User 1000008 <1000008@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0