Update patch set 1
Patch Set 1:

There is a bit of discussion of this at:
http://lists.infradead.org/pipermail/linux-arm-kernel/2014-April/244356.html

Based on the description of 761320 - I'm really surprised enabling the workaround for it is fixing the issue, since the errata is a deadlock condition involving cache, not anything related to the FP registers.  I'm wondering if it is just luck that you haven't had a fuse failure when running this?

Patch-set: 1