Update patch set 1
Patch Set 1:

Sorry, forgot to insert this link: http://forum.xda-developers.com/showpost.php?p=55827815&postcount=592

Yes, it doesn't sound like this should solve it. But it either does, or hides some of it symptoms.

I've checked a lot of Samsung source drops, and as I say in the xda post, they never seem to have it on for real 4210 source drops. Except in the S4 i9500 kernel there's a remnant of exynos4 which does have it on for 4210 as well...

Patch-set: 1