Update patch set 1
Patch Set 1:

Further testing shows that when combining this with other changes the bug shows up again. So that proves that the root cause is not solved with this; just some timing/memory layout/whatever changed so it does not occur or is less likely to. Maybe an idea for smdk4412 and smdk4212 to disable this errata and check for performance improvement?

Anyway, Samsung does also enable this for 4212, indirectly, because all 4212 devices also get the 4412 settings...

Patch-set: 1